First pass at moving per round checkpointing to basic client
