+++ research/picai/fedavg/server.py
@@ -10,11 +10,11 @@ from flwr.server.client_manager import SimpleClientManager
 from flwr.server.strategy import FedAvg
 
 from fl4health.parameter_exchange.full_exchanger import FullParameterExchanger
+from fl4health.server.base_server import FlServerWithCheckpointing
 from fl4health.utils.config import load_config
 from fl4health.utils.metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 from fl4health.utils.parameter_extraction import get_all_model_parameters
 from research.picai.model_utils import get_model
-from research.picai.picai_server import PicaiServer
 
 
 def fit_config(
@@ -63,7 +63,7 @@ def main(config: Dict[str, Any], server_address: str, n_clients: int) -> None:
         initial_parameters=get_all_model_parameters(model),
     )
 
-    server = PicaiServer(
+    server = FlServerWithCheckpointing(
         client_manager=client_manager,
         model=model,
         parameter_exchanger=FullParameterExchanger(),
