+++ research/picai/fl_nnunet/config.yaml
@@ -1,12 +1,10 @@
 # FL4Health Settings:
 n_clients: 1
-n_server_rounds: 1
+n_server_rounds: 1000
 server_address: '0.0.0.0:8080'
-local_epochs: 1
-# local_steps: 20
-# starting_checkpoint: /home/shawn/Code/nnunet_storage/nnUNet_results/Dataset004_Hippocampus/nnUNetTrainer_1epoch__nnUNetPlans__2d/fold_0/checkpoint_best.pth
+local_steps: 256
 
 # nnunet Settings:
 nnunet_config: 2d
 fold: 0
-#nnunet_plans: /home/shawn/Code/nnunet_storage/nnUNet_preprocessed/Dataset004_Hippocampus/nnUNetPlans.json
+nnunet_plans: /ssd003/projects/aieng/public/PICAI/nnUNet/nnUNet_preprocessed/Dataset005_Haider_710/nnUNetPlans.json
